requirements.txt
sqlalchemy==1.3.20
passlib==1.7.4
pymysql==0.10.1
pyjwt==1.7.1
Flask==1.1.2
cryptography==3.2
mysqlclient==2.0.1
